Floating IP assignment to support specific port
When creating VMs with more than one port, the floating IP assignment
can randomly choose one of the created ports, potentially leading to
failures.

This patch aims to provide a way to specify which port to assign the
floating IP to. The network where the port lives must be specified
thus the assignment becomes deterministic.

This patch should keep compatibility with previous versions of the
neutron_floating_ip module.

Signed-off-by: Abel Navarro <abel@midokura.com>